# Conflicts:
#	FusorCompModeling/nbproject/private/private.properties
#	FusorVis/build/built-jar.properties
#	FusorVis/build/classes/fusorvis/FusorVis$1.class
#	FusorVis/build/classes/fusorvis/FusorVis$2.class
#	FusorVis/build/classes/fusorvis/FusorVis$3.class
#	FusorVis/build/classes/fusorvis/FusorVis$4.class
#	FusorVis/build/classes/fusorvis/FusorVis.class
#	FusorVis/dist/FusorVis.html
#	FusorVis/dist/FusorVis.jar
#	FusorVis/dist/FusorVis.jnlp
#	FusorVis/dist/lib/FusorCompModeling.jar
#	FusorVis/dist/lib/MathEval.jar
#	FusorVis/nbproject/private/private.properties
#	FusorVis/nbproject/project.properties
#	FusorVis/src/fusorvis/FusorVis.java

+++ FusorCompModeling/src/fusorcompmodeling/Acceleration.java
@@ -13,7 +13,7 @@ public class Acceleration {
     public double accX;
     public double accY;
     public double accZ;
-    public Acceleration(int accx, int accy, int accz){
+    public Acceleration(double accx, double accy, double accz){
         accX = accx;
         accY = accy;
         accZ = accz;
